# TennisPlatformApp/openapi.yaml
openapi: 3.0.3
info:
  title: Tennis Platform API
  version: "0.1.0-draft"
  description: >
    Contrato de API del MVP de Tennis Platform. Borrador de diseño (ver
    Documentos/_arquitectura/11-contrato-api.md para las convenciones generales
    y el mapeo de códigos de negocio a HTTP).
servers:
  - url: http://localhost:8080/api/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Registro público de alumno
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "202":
          description: >
            Siempre 202 y sin cuerpo, exista o no una cuenta con ese email. Devolver 201
            cuando se crea y 409 cuando ya existe permitiría enumerar usuarios, que es
            precisamente lo que prohíbe 08-security-engineer.md. Si el email ya estaba
            registrado no se crea nada y se avisa por correo al titular real de la dirección.
        "400": { $ref: "#/components/responses/ValidationError" }
        "429":
          description: Demasiados intentos desde la misma IP
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Login correcto. El refresh token se fija en una cookie HttpOnly (`Set-Cookie`).
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=...; HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: >
            Único resultado de cualquier fallo: email inexistente, contraseña incorrecta o
            cuenta desactivada. Distinguirlos revelaría qué direcciones tienen cuenta.
            Una cuenta pendiente de verificación SÍ puede iniciar sesión: la verificación
            condiciona reservar, no acceder (ver 13-fase5-analisis-identity.md, decisión 8.5).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "429":
          description: Demasiados intentos desde la misma IP
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout (revoca el refresh token actual)
      responses:
        "204": { description: Sesión cerrada }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rota el access token usando el refresh token de la cookie
      security: []
      responses:
        "200":
          description: Nuevo access token emitido; refresh token rotado (nueva cookie)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccessTokenResponse" }
        "401":
          description: Refresh token ausente, expirado, revocado o reusado (familia revocada)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Verifica el email con el token recibido por correo
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyEmailRequest" }
      responses:
        "204": { description: Email verificado }
        "400": { $ref: "#/components/responses/ValidationError" }
        "409":
          description: Token inválido, expirado o ya usado
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Solicita un email de recuperación de contraseña
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotPasswordRequest" }
      responses:
        "202":
          description: >
            Siempre 202, exista o no la cuenta, para no permitir enumeración
            de usuarios (ver 08-security-engineer.md).
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Restablece la contraseña con el token recibido por correo
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
      responses:
        "204": { description: Contraseña actualizada }
        "400": { $ref: "#/components/responses/ValidationError" }
        "409":
          description: Token inválido, expirado o ya usado
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: No autenticado o token expirado
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
    ValidationError:
      description: Formato o validación de campos incorrecta
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
    ProblemDetails:
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: Código de negocio, ver 11-contrato-api.md
          example: LESSON_FULL
        instance: { type: string }
    Role:
      type: string
      enum: [ADMIN, TEACHER, STUDENT]
    UserStatus:
      type: string
      enum: [PENDING_VERIFICATION, ACTIVE, DISABLED]
...
    UserSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        status: { $ref: "#/components/schemas/UserStatus" }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 10 }
      description: >
        fullName se retira hasta la Fase 6: el nombre vive en student_profiles, tabla que
        todavía no existe. Aceptar el campo y descartarlo haría creer al cliente que se ha
        guardado. Volverá cuando el módulo student pueda persistirlo.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/UserSummary" }
    AccessTokenResponse:
      type: object
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token: { type: string }
        newPassword: { type: string, format: password, minLength: 10 }
...